requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.22.0 
pandas
matplotlib
seaborn
scikit-learn
paho-mqtt